+++ src/main/kotlin/sh/zachwal/button/auth/CurrentUser.kt
@@ -6,11 +6,11 @@ import io.ktor.response.respond
 import io.ktor.sessions.get
 import io.ktor.sessions.sessions
 import sh.zachwal.button.db.jdbi.User
-import sh.zachwal.button.session.SessionPrincipal
+import sh.zachwal.button.session.principals.UserSessionPrincipal
 import sh.zachwal.button.users.UserService
 
 suspend fun currentUser(call: ApplicationCall, userService: UserService): User {
-    val p = call.sessions.get<SessionPrincipal>()
+    val p = call.sessions.get<UserSessionPrincipal>()
     return p?.let { userService.getUser(p.user) } ?: run {
         call.respond(HttpStatusCode.Unauthorized)
         throw UnauthorizedException()
