+++ src/main/kotlin/sh/zachwal/button/auth/SessionAuth.kt
@@ -5,17 +5,17 @@ import io.ktor.auth.Authentication
 import io.ktor.auth.session
 import io.ktor.response.respondRedirect
 import org.slf4j.LoggerFactory
-import sh.zachwal.button.session.SessionPrincipal
+import sh.zachwal.button.session.principals.UserSessionPrincipal
 
 private val logger = LoggerFactory.getLogger("SessionAuth")
 
 fun Authentication.Configuration.configureSessionAuth() {
-    session<SessionPrincipal> {
+    session<UserSessionPrincipal> {
         challenge {
             call.respondRedirect("/login")
         }
         validate {
-            it.takeIf(SessionPrincipal::isValid)
+            it.takeIf(UserSessionPrincipal::isValid)
         }
     }
 }
